feat: moved honeypot helper functions to helpers.py
